+++ src/utils/Bilibili/bilifavOperate.ts
@@ -2,7 +2,7 @@ import { humanishApiLimiter } from '@APM/utils/mediafetch/throttle';
 import { getFavListBVID } from '@APM/utils/mediafetch/bilifavlist';
 import { getCookie, sendBVFavorite } from '../BiliOperate';
 import { PlayListDict } from '../ChromeStorage';
-import { getBiliUser } from '../PersonalCloudAuth';
+import { getBiliUser } from '../sync/PersonalCloudAuth';
 
 const BILI_GETFAVLIST_API =
   'https://api.bilibili.com/x/v3/fav/folder/created/list-all?up_mid={mid}';
